+++ lib/db/schema.ts
@@ -180,7 +180,7 @@ export const column = pgTable(
     id: bigint("id", { mode: "number" }).primaryKey().generatedAlwaysAsIdentity(),
     created_at: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
     type: text("type", {
-      enum: ["url", "text", "image", "channel", "pdf", "video", "tweet", "youtube"],
+      enum: ["url", "text", "image", "channel", "pdf", "video", "tweet", "youtube", "spotify"],
     }).notNull(),
     title: text("title"),
     description: text("description"),